Per-model schema selectors
Still need to think about invalidation of schema and attribute lists
attached to attribute fields when schemata are saved.  The schema cache
is invalidated when schemata are saved, but the references from objects
containing attribute fields aren't.  The handling of multiple schemata
also needs a look in the example app.